Extract core from controller
